+++ src/client/hooks/api/useDataApi.test.tsx
@@ -2,7 +2,7 @@ import { act, renderHook, waitFor } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 
 import { getDefaultState, useDataApi } from './useDataApi';
-import { bffApiHost } from '../../../testing/setupTests';
+import { bffApiHost } from '../../../testing/setup';
 import { bffApi } from '../../../testing/test-utils';
 
 interface ResponseData {
