+++ src/server/helpers/source-api-request.test.ts
@@ -18,7 +18,7 @@ import {
   getRequestConfigCacheKey,
   requestData,
 } from './source-api-request';
-import { remoteApiHost } from '../../testing/setupTests';
+import { remoteApiHost } from '../../testing/setup';
 import { remoteApi } from '../../testing/test-utils';
 import {
   apiErrorResult,
