+++ src/server/services/buurt/datasets.test.ts
@@ -6,7 +6,7 @@ import {
   transformHardlooproutesResponse,
   transformMeldingenBuurtResponse,
 } from './datasets';
-import { remoteApiHost } from '../../../testing/setupTests';
+import { remoteApiHost } from '../../../testing/setup';
 
 describe('Custom dataset tranformations', () => {
   it('Should group distance of hardlooproute', () => {
