+++ src/server/services/simple-connect/api-service.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, test } from 'vitest';
 
 import * as service from './api-service';
-import { remoteApiHost } from '../../../testing/setupTests';
+import { remoteApiHost } from '../../../testing/setup';
 import { remoteApi } from '../../../testing/test-utils';
 
 const REQUEST_ID = 'test-x';
