+++ src/testing/test-utils.ts
@@ -2,7 +2,7 @@ import { Request, Response } from 'express';
 import nock from 'nock';
 import UID from 'uid-safe';
 
-import { bffApiHost, remoteApiHost } from './setupTests';
+import { bffApiHost, remoteApiHost } from './setup';
 import { AuthProfile, AuthProfileAndToken } from '../server/auth/auth-types';
 import { createOIDCStub } from '../server/routing/router-development';
 import { HTTP_STATUS_CODES } from '../universal/constants/errorCodes';
